Add webhook to validate STS configurations at pull request.
Since we don't give any information at request time about why a STS
policy is valid or not so that we don't leak details about the policy,
this webhook will validate on the repo at pull request / push so the
repo authors get feedback on potential config issues.

This changes does not set up GCLB, since this probably needs some more
discussion for how we set this up to not conflict with the existing STS
service.