Add option to prefer fetching policies from org repo.
This adds the ability to have a server-side flag to allow configuring
OctoSTS to only read policies from the org/.github repo. This allows for
policy centralization for deployments that prefer stricter
consolidation and enforcement.

This also adds a test harness for the exchange flow to make it easier to
test and see what values are being exchanged.

Co-authored-by: Karl Haworth <karl.haworth@aa.com>