docs: adds octo sts documentation
It would take quite a while for me to read this repository entirely to
grok what the contents as well as some of the purpose of the code. For
this reason I have used Claude Code to analyse this repository to act as
a technical writer in attempts to document what it finds in this
repository to help those who want to contribute/self-host.

Signed-off-by: ChrisJBurns <29541485+ChrisJBurns@users.noreply.github.com>