Exchange: Return InvalidArgument if scope or identity are missing (#1001)
Right now if you pass in an empty scope, you get the cryptic error:

```
no installation found for "."
```

Because we try to dir parse "" and try to look up the org "." which
never works. https://go.dev/play/p/zxiNbeV-gG5

This changes the behavior to return a better error message and to not
bother trying to search for matching installations if this happens.